requirements.txt
tweepy
schedule
openai
pyhton-dotenv